api/requirements.txt
python-telegram-bot==13.7
redis==4.5.4
celery==5.3.1
django-celery-results==2.5.1
python-dotenv==1.0.0